Reorganize Rest starter, fix some issues with sample app
