+++ easyms-training-rest-starter/src/main/java/com/easyms/training/rest/autoconfigure/EasyMsTrainingAutoConfiguration.java
@@ -1,4 +1,4 @@
-package com.easyms.training.common;
+package com.easyms.training.rest.autoconfigure;
 
 import lombok.Data;
 import lombok.extern.slf4j.Slf4j;
@@ -8,7 +8,7 @@ import org.springframework.context.annotation.Configuration;
 @Slf4j
 @Data
 @Configuration
-@ComponentScan(basePackages = {"com.easyms.training.common"})
+@ComponentScan(basePackages = {"com.easyms.training.rest.ms"})
 public class EasyMsTrainingAutoConfiguration {
 
 }
