+++ easyms-training-rest-starter/src/main/java/com/easyms/training/rest/ms/config/SwaggerProperties.java
@@ -1,4 +1,4 @@
-package com.easyms.training.common.config;
+package com.easyms.training.rest.ms.config;
 
 import lombok.Data;
 import org.springframework.boot.context.properties.ConfigurationProperties;
