+++ pom.xml
@@ -23,6 +23,7 @@
     <properties>
         <maven.version>3.5.0</maven.version>
         <easyms.sdk.version>1.0.0-SNAPSHOT</easyms.sdk.version>
+        <swagger.version>2.9.2</swagger.version>
     </properties>
 
     <modules>
@@ -30,6 +31,25 @@
         <module>easyms-training-rest-starter</module>
     </modules>
 
+    <dependencyManagement>
+        <dependencies>
+            <dependency>
+                <groupId>io.springfox</groupId>
+                <artifactId>springfox-swagger2</artifactId>
+                <version>${swagger.version}</version>
+            </dependency>
+            <dependency>
+                <groupId>io.springfox</groupId>
+                <artifactId>springfox-swagger-ui</artifactId>
+                <version>${swagger.version}</version>
+            </dependency>
+            <dependency>
+                <groupId>com.easyms</groupId>
+                <artifactId>easyms-training-rest-starter</artifactId>
+                <version>1.0.0-SNAPSHOT</version>
+            </dependency>
+        </dependencies>
+    </dependencyManagement>
     <dependencies>
         <dependency>
             <groupId>org.projectlombok</groupId>
